requirements.txt
amqp==5.0.9
anyjson==0.3.3
APScheduler==3.8.1
asgiref==3.4.1
asyncio==3.4.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.20.32
botocore==1.23.32
cached-property==1.5.2
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.11
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.11.0
django-email==0.1.10
django-on-heroku==1.1.2
django-storages==1.12.3
django-telethon-session==0.9
django-timezone-field==4.2.3
djangorestframework==3.13.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
jmespath==0.10.0
kombu==5.2.3
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
pyaes==1.6.1
pyasn1==0.4.8
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.5
python-http-client==3.3.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.1
redis-server==6.0.9
requests==2.27.1
rsa==4.8
s3transfer==0.5.0
sendgrid==6.9.5
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
Telethon==1.24.0
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.7.0
